web/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dependency-injector==4.42.0
Django==5.1.2
django-celery-results==2.5.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
packaging==24.2
pillow==11.0.0
postgres==4.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.16.0